feat: Add a get_hidden_directory function in the file module.

+++ discover_test_coverage/file.py
@@ -18,11 +18,30 @@ def find_python_files(directory: Path) -> List[Path]:
     # with that said, this works by a convention that a developer
     # may not follow (i.e., tests could be inside program directories)
     all_python_files = sorted(directory.rglob(constants.wildcards.All_Python))
-    output.logger.debug(f"Type of all_python_files: {type(all_python_files)}")
     return all_python_files
 
 
+def find_python_files_not_conftest(directory: Path) -> List[Path]:
+    """Determine if the list of Path objects contains one for the conftest file."""
+    not_conftest_file_list = []
+    python_file_list = find_python_files(directory)
+    for python_file in python_file_list:
+        if python_file.name != "conftest.py":
+            not_conftest_file_list.append(python_file)
+    return not_conftest_file_list
+
+
 def find_conftest_files(directory: Path) -> List[Path]:
+    """Determine if the list of Path objects contains one for the conftest file."""
+    conftest_file_list = []
+    python_file_list = find_python_files(directory)
+    for python_file in python_file_list:
+        if python_file.name == "conftest.py":
+            conftest_file_list.append(python_file)
+    return conftest_file_list
+
+
+def find_and_create_conftest_files(directory: Path) -> List[Path]:
     """Determine if the list of Path objects contains one for the conftest file."""
     conftest_file_list = []
     python_file_list = find_python_files(directory)
@@ -49,7 +68,9 @@ def elide_path(path, maximum_parts: int = 4, include_up_to: int = 5):
     return PurePath(*parts)
 
 
-def create_hidden_directory(containing_directory: Path, directory: Path) -> Path:
+def create_hidden_directory(
+    containing_directory: Path, directory: Path, delete_path: bool = True
+) -> Path:
     """Create a hidden directory."""
     output.logger.debug(f"Hiding in containing directory: {containing_directory}")
     output.logger.debug(f"Hiding the directory with name: {directory.name}")
@@ -59,11 +80,26 @@ def create_hidden_directory(containing_directory: Path, directory: Path) -> Path
         containing_directory / Path(constants.markers.Hidden + str(directory.name))
     )
     # delete the hidden directory if it already exists
-    if hidden_directory.exists() and hidden_directory.is_dir():
+    if hidden_directory.exists() and hidden_directory.is_dir() and delete_path:
         rmtree(hidden_directory)
     # create the hidden directory for storing instrumentation
     # note that mkdir throws an exception if the directory exists;
     # since it was already deleted, the exception should not occur
     hidden_directory.mkdir(parents=True)
     output.logger.debug(f"Created the hidden directory: {hidden_directory}")
+    return hidden_directory
+
+
+def get_hidden_directory(
+    containing_directory: Path, directory: Path, delete_path: bool = True
+) -> Path:
+    """Create a hidden directory."""
+    output.logger.debug(f"Hiding in containing directory: {containing_directory}")
+    output.logger.debug(f"Hiding the directory with name: {directory.name}")
+    # create a hidden directory (i.e., a directory with name "program" will
+    # lead to creation of a hidden directory called ".program")
+    hidden_directory = Path(
+        containing_directory / Path(constants.markers.Hidden + str(directory.name))
+    )
+    output.logger.debug(f"Accessing the hidden directory: {hidden_directory}")
     return hidden_directory
